pyproject.toml
[tool.poetry]
name = "fitproject"
version = "0.1.0"
description = ""
authors = ["Agata Tekien <tekienagata@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"
docker = "^7.1.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
